BitKeeper/etc/ignore:
  auto-union
extra/comp_err.c:
  Auto merged
include/my_pthread.h:
  Auto merged
mysql-test/r/innodb.result:
  Auto merged
mysql-test/r/mix2_myisam.result:
  Auto merged
mysql-test/r/mysqltest.result:
  Auto merged
mysql-test/r/type_blob.result:
  Auto merged
mysql-test/t/disabled.def:
  Auto merged
mysql-test/t/mysqltest.test:
  Auto merged
sql/Makefile.am:
  Auto merged
sql/item.cc:
  Auto merged
sql/item_cmpfunc.cc:
  Auto merged
sql/item_cmpfunc.h:
  Auto merged
sql/item_strfunc.cc:
  Auto merged
sql/log.cc:
  Auto merged
sql/mysql_priv.h:
  Auto merged
sql/set_var.cc:
  Auto merged
sql/sql_class.cc:
  Auto merged
sql/sql_class.h:
  Auto merged
sql/sql_parse.cc:
  Auto merged
sql/sql_select.cc:
  Auto merged
sql/tztime.cc:
  Auto merged
client/mysqltest.c:
  Manual merge
mysql-test/mysql-test-run.pl:
  Manual merge
sql/mysqld.cc:
  Manual merge

+++ storage/innobase/dict/dict0load.c
@@ -843,7 +843,7 @@ err_exit:
 	table = dict_mem_table_create(name, space, n_cols & ~0x80000000UL,
 				      flags);
 
-	table->ibd_file_missing = ibd_file_missing;
+	table->ibd_file_missing = (unsigned int) ibd_file_missing;
 
 	ut_a(name_of_col_is(sys_tables, sys_index, 3, "ID"));
 
@@ -1180,8 +1180,8 @@ dict_load_foreign(
 
 	/* We store the type in the bits 24..29 of n_fields_and_type. */
 
-	foreign->type = n_fields_and_type >> 24;
-	foreign->n_fields = n_fields_and_type & 0x3FFUL;
+	foreign->type = (unsigned int) (n_fields_and_type >> 24);
+	foreign->n_fields = (unsigned int) (n_fields_and_type & 0x3FFUL);
 
 	foreign->id = mem_heap_strdup(foreign->heap, id);
